Add base class for piece rendering
Added a base class that handles the graphics of the Piece class. Also added another argument, surface, to be used to center the piece in the middle of the square (which requires a surface). Added a class variable dictionary for directions on the board (for pieces) and methods to get the forward directon, left-right, etc. Added a decorator that automatically highlights the valid squares a piece can move to.