Method for moving pieces and pawn move generation fixed
Wrote a simple helper method that refactors a bit of code for piece moving; it moves the piece and then centers on the screen. Also fixed an error in move generation for the pawn. Previously, a pawn could be moved two squares up even if there was a piece blocking it, but with this commit, that has been fixed.