Use a dictionary to get piece on a square
Made a change in the algorithm for getting a piece on a square. Previously, I was using a list to store pieces (which I still am) and then looping through the whole list to get a piece on a square. But now, I'm making use of the almost constant time lookup functionality of a dictionary to speed the algorithm up by 1.5 ten thousandths of a second.