Use spritesheet to load piece images
Implemented the usage of a spritesheet that contains the images of the pieces. This is a much faster method of loading/getting the required image for a piece.

Added a PieceCreator class, which could also be turned into a 'static' class.

Deleted single piece image files.

Deleted the 'sounds' and 'pieces' directories, which contained multiple files (especially 'pieces') but since those files have been replaced with a single 'spritesheet', it was no longer necessary to have such a folder, as they are all under the same 'assets' directory.